refactor: implement dfd v2 support as well as new package management

+++ src/wasm/src/util.rs
@@ -1,6 +1,15 @@
-use std::{fs, io, path::Path};
+use std::{
+    error::Error,
+    fs,
+    io::{self},
+    path::Path,
+};
 
-use navigation_database::util::{get_path_type, PathType};
+use navigation_database::{
+    traits::InstalledNavigationDataCycleInfo,
+    util::{get_path_type, PathType},
+};
+use uuid::Uuid;
 
 pub fn path_exists(path: &Path) -> bool {
     get_path_type(path) != PathType::DoesNotExist
@@ -26,13 +35,21 @@ pub fn delete_folder_recursively(path: &Path, batch_size: Option<usize>) -> io::
         let path = entry.path();
         let path_type = get_path_type(&path);
 
+        if path.file_name().unwrap() == "" {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        }
+
         if path_type == PathType::Directory {
             delete_folder_recursively(&path, batch_size)?;
         } else if path_type == PathType::File {
             fs::remove_file(&path)?;
-        } else if let None = path.extension() {
-            // There are edge cases where completely empty directories are created and can't be deleted. They get registered as "unknown" path type so we need to check if the path has an extension (which would tell us if it's a file or a directory), and if it doesn't, we delete it as a directory
-            let _ = fs::remove_dir(&path); // this can fail silently, but we don't care since there also might be cases where a file literally doesn't exist
+        } else if path.extension().is_none() {
+            // There are edge cases where completely empty directories are created and can't be deleted. They get
+            // registered as "unknown" path type so we need to check if the path has an extension (which would tell us
+            // if it's a file or a directory), and if it doesn't, we delete it as a directory
+            let _ = fs::remove_dir(&path); // this can fail silently, but we don't care since there also might be cases
+                                           // where a file literally doesn't exist
         }
     }
     // Check if the directory is empty. If it is, delete it
@@ -59,11 +76,22 @@ pub fn copy_files_to_folder(from: &Path, to: &Path) -> io::Result<()> {
     // Create the directory we are copying to
     fs::create_dir(to)?;
     // Collect the entries that we will copy
-    let entries = fs::read_dir(from)?.collect::<Result<Vec<_>, _>>()?;
+    let entries = fs::read_dir(from)?;
+
     // Copy the entries
     for entry in entries {
+        let Ok(entry) = entry else {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        };
+
         let path = entry.path();
         let path_type = get_path_type(&path);
+
+        if path.file_name().unwrap() == "" {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        }
 
         if path_type == PathType::Directory {
             let new_dir = to.join(path.file_name().unwrap());
@@ -79,4 +107,26 @@ pub fn copy_files_to_folder(from: &Path, to: &Path) -> io::Result<()> {
 
 pub fn trim_null_terminator(s: &str) -> &str {
     s.trim_end_matches(char::from(0))
+}
+
+pub fn generate_uuid_from_path<P>(cycle_path: P) -> Result<String, Box<dyn Error>>
+where
+    P: AsRef<Path>,
+{
+    let cycle = &serde_json::from_reader(fs::File::open(cycle_path)?)?;
+
+    Ok(generate_uuid_from_cycle(cycle))
+}
+
+pub fn generate_uuid_from_cycle(cycle: &InstalledNavigationDataCycleInfo) -> String {
+    let uuid_hash = format!(
+        "{}{}{}{}",
+        cycle.cycle, cycle.revision, cycle.format, cycle.name
+    );
+
+    let uuid_uuid = Uuid::new_v3(&Uuid::NAMESPACE_URL, uuid_hash.as_bytes());
+
+    let uuid_hypenated = uuid_uuid.as_hyphenated();
+
+    uuid_hypenated.to_string()
 }
